ide: possible OOB access into sOperatorTab with signed-char index -128..-1
Seems the table is defined for 0..255 range, but `*s` is signed char on
some platforms.

+++ uppsrc/ide/clang/Signature.cpp
@@ -330,8 +330,8 @@ Vector<ItemTextPart> ParsePretty(const String& name, const String& signature, in
 				s++;
 		}
 		else
-		if(sOperatorTab[*s]) {
-			while(sOperatorTab[s[n]])
+		if(sOperatorTab[uint8(*s)]) {
+			while(sOperatorTab[uint8(s[n])])
 				n++;
 			p.type = ITEM_CPP;
 		}
